requirements/base.txt
Pillow==4.3.0
psycopg2==2.6.2
python-memcached==1.58
Jinja2==2.9.5
Django==2.0

easy-thumbnails==2.5
django-image-cropping==1.0.4

django-allauth==0.30.0
-e git://github.com/django-ckeditor/django-ckeditor.git#egg=master
# django-ckeditor==5.3.1
